feat: Move AuthViewModel to shared usecase module (Phase 30)
Move AuthViewModel interface and DefaultAuthViewModel to usecase/commonMain,
completing the ViewModel extraction. All ViewModels now live in shared KMP
modules. Also fix pre-existing FQN violations in LogSummaryCard.
